refactor: remove observation pipeline and embedding infrastructure
This massive refactoring removes the entire observation write path
(observe -> compress -> dedup -> embed -> hybrid search) and all
embedding-related code from tidev. The memory system is now purely
based on FTS5 keyword search and session summary LLM calls.

## Removed modules (5 files deleted):
- src/memory/observe.rs
- src/memory/compress.rs
- src/memory/compression_queue.rs
- src/memory/dedup.rs
- src/memory/hybrid_search.rs

## Removed types:
HookPayload, HookType, ObservationType, RawObservation,
CompressedObservation, ObservationResult, Modality,
HybridSearchResult, HybridScore, Bm25Index,
EmbeddingActiveModel, EmbeddingModelSummary, EmbeddingModelConfig,
BackgroundTasks

## Removed MemoryStore fields (10):
dedup, bm25, compression_enabled, llm_compression, compression_model,
embedding_model, hybrid_search, compressor_tx, compression_cb_failures,
compression_cb_tripped_at

## Removed ~20 MemoryStore methods:
observe, compress, schedule_compression, schedule_embedding_backfill,
recover_uncompressed, backfill_embeddings, set_compression_*,
set_llm_compression, set_compression_sender, set_embedding_model,
resolve_compression_llm, load_recent_compressed_observations,
list_recent_observations, graph_extract_from_observation,
search_hybrid_with, search_hybrid

## Removed config fields:
memory.compression_enabled, memory.llm_compression,
memory.compression_model, memory.embedding_model

## Removed dependency:
sqlite-vec

## Rewritten consumer modules:
- sessions.rs: summarize_session reads from messages table (zstd
  decompressed) instead of compressed_observations, preserving
  append-only constraint for provider prefix caching
- patterns.rs: mine_co_change reads tool result messages (file_diffs +
  content); mine_error_repeats reads tool result error patterns
- graph.rs: added extract_from_session_summary /
  extract_from_memory_entry, integrated into consolidation pipeline

## Storage layer:
- Added load_session_messages() standalone function
- Added FTS5 startup rebuild + incremental write on remember()
- Removed compressed_observations, vec_obs_map from schema

## TUI/UI:
- Removed observations panel, compression/embedding toggles
- Removed EmbeddingModel variant from ModelPanelItem
- Model panel sidebar simplified to single Summarization entry

## All callers cleaned:
hooks/engine.rs, agent/runtime.rs, tui/core/run.rs, web/mod.rs,
gateway/mod.rs, config/mod.rs, provider_setup/, web routes/,
presets.toml

244 tests pass, zero new warnings.

+++ src/memory/evict.rs
@@ -21,7 +21,7 @@ impl EvictionService {
             )",
             [],
         )?;
-        report.stale_memories_removed = stale as usize;
+        report.stale_memories_removed = stale;
 
         // 2. Old non-latest versions
         let old_versions = db.execute(
@@ -29,7 +29,7 @@ impl EvictionService {
              AND julianday('now') - julianday(updated_at) > 30",
             [],
         )?;
-        report.old_versions_removed = old_versions as usize;
+        report.old_versions_removed = old_versions;
 
         // 3. Clean up retention_scores for deleted entities
         db.execute(
